removing Online Compiler. Moved ocloc_api.h to KernelCompiler directory (needed for compiling OpenCL C). Still needs Win ABI symbols
